tests/wpt/web-platform-tests/css/work-in-progress/ttwf_bj/Chris/ttwf-scrollintoview.html

<!DOCTYPE html>
<html>
<head>
    <title>CSSOM View - 6 - scrollIntoView tests</title>
    <meta charset="utf-8">
    <link rel="author" title="Chris Wu" href="mailto:pwx.frontend@gmail.com">
	<link rel="help" href="https://drafts.csswg.org/cssom-view/#widl-Element-scrollIntoView-void-boolean-top">
    <meta name="flags" content="dom">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style type="text/css">
    	#phitem{height: 2000px;}
    	#viewitem{width:300px;height:200px;margin:0 auto;background-color: lightgreen;}
    </style>  
</head>
<body>
    <div id="myDiv">
    	<div id="phitem"></div>
    	<div id="viewitem"></div>
    </div>
    <div id="log"></div>
    <script>
		var asytest = async_test('check scrollIntoView');
		asytest.step(function(){
			var viewitem = document.getElementById('viewitem');
			viewitem.scrollIntoView();
			assert_true(window.pageYOffset + window.innerHeight > viewitem.offsetTop, "The element isn't in the viewport")
		})
		asytest.done()
    </script>    
</body>
</html>
